requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
aiohttp>=3.9.5
pysellauth>=0.1.1
aiosqlite>=0.20.0
